split ApolloClient and fragments out of index.ts
