+++ test/ObservableQuery.ts
@@ -7,7 +7,7 @@ import gql from 'graphql-tag';
 import mockQueryManager from './mocks/mockQueryManager';
 import mockWatchQuery from './mocks/mockWatchQuery';
 import { ObservableQuery } from '../src/ObservableQuery';
-import { ApolloQueryResult } from '../src';
+import { ApolloQueryResult } from '../src/queryManager';
 
 // I'm not sure why mocha doesn't provide something like this, you can't
 // always use promises
